Changes to Problems 1-2
I changed the number of samples from "n*2" to "n" for options 4 and 5,
when we're sampling from multivariate normal distributions. The reason
for this is because multivariate normal random samples are vectors equal
to the number of parameters (in this case 2), so by multiplying it by 2
we were getting twice as many observations as we wanted.
I also changed the way we sampled for option=5; the old way might have
been mixing up results from different columns of the multivariate
distribution.
Finally, I changed the output of the function from have five different
potential outputs (e.g. voter1, voter2, etc.) to all be labeled "voter";
I don't think this should cause any problems since we're only generating
one matrix at a time.