Updated Problem 3
Had a problem with the following line of code in problem 3:
apply((voter.position - party.1)^2, 1, sum)

I'm not sure this was working correctly. When I tried subtracting
voter.position and party.1, it would subtract it in the reverse order
that we wanted. (i.e. instead of subtracting the two values of party.1
across the rows like we wanted, it was subtracting down the columns...so
the first column of voter.positions was sometimes having the first value
of party.1 being subtracted from it, and other times it was having the
second value of party.1 being subtracted from it.)

To be safe I decided to calculate the difference in positions as a
matrix first, then go through and add them up. This may have been
unnecessary, in which case I'm sorry for making the code more complex. I
can try and describe in better detail what I think was wrong with the
old code if you want, just let me know.