Initial Attempt at Problem 5
Not quite done yet; still need to make a few changes, but I should have
the "master" function described in the homework done soon. At the moment
the main problem seems to be that my visualization2 function is
recalculating the voter preferences when I run it through the for loop.
I need to go back and change how this function (and affiliation2) work
so that they do not create new preferences every time.